TOOLS-100 replace deployService function
